Adding caching for Hjerson

+++ libs/Metrics/Hjerson.php
@@ -1,5 +1,6 @@
 <?php
-
+use Nette\Caching\Cache;
+use Nette\Caching\Storages\FileStorage;
 /**
  * Hjerson external metric implementation
  */
@@ -8,10 +9,13 @@ class Hjerson implements IMetric {
 	private $referenceText = array();
 	private $translationText = array();
 	private $type = null;
+	private $cache = null;
 	private $externalCmd = "python external/hjerson+.py";
 
 	public function __construct($type = "") {
 		$this->type = $type;
+        $storage = new FileStorage( "storage/hjerson" );
+        $this->cache = new Cache($storage  );
 	}
 
 	public function init() {
@@ -26,6 +30,16 @@ class Hjerson implements IMetric {
 		return 0;
 	}
 
+    private function processOutput($output){
+              foreach ($output as $line){
+                  $tempArray = explode("\t", $line);
+                  $typeName = trim(str_replace(":","",$tempArray[0]));
+                  if ($typeName == $this->type){
+                      return $tempArray[1]; // if you want to use % representation, use [2] instead
+                  }
+              }      
+    }
+
 	public function getScore() {
 		$hashId = md5(time());
 		$reference = "temp/ref." . $hashId;
@@ -33,37 +47,24 @@ class Hjerson implements IMetric {
 		$refTxt = implode("\n", $this->referenceText);
 		$hypTxt = implode("\n", $this->translationText);
 		if ($refTxt != "" && $hypTxt != ""){
-            $textsHash = md5('Hyp:' . $hypTxt . "Ref:" . $refTxt);
-            $temporaryResultsFile = "temp/hjerson." . $textsHash;
-            if (file_exists($temporaryResultsFile)){
-              $output = file($temporaryResultsFile);
-              foreach ($output as $line){
-                  $tempArray = explode("\t", $line);
-                  $typeName = trim(str_replace(":","",$tempArray[0]));
-                  if ($typeName == $this->type){
-                      return $tempArray[1]; // if you want to use % representation, use [2] instead
-                  }
+            $key = md5('Hyp:' . $hypTxt . "Ref:" . $refTxt);
+            if( $this->cache->load( $key ) !== NULL){
+              $output = explode("\n", $this->cache->load( $key ));
+            } else {
+              file_put_contents($reference, $refTxt);
+              file_put_contents($hypothesis, $hypTxt);
+              $cmd = sprintf("%s -R %s -H %s", $this->externalCmd, $reference, $hypothesis);
+              exec($cmd, $output, $return);
+              unlink($reference);
+              unlink($hypothesis);
+              if ($return != 0){
+                  return -1;
               }
+              $outputToStore = implode("\n", $output);
+              $this->cache->save($key, $outputToStore);
             }
-			file_put_contents($reference, $refTxt);
-			file_put_contents($hypothesis, $hypTxt);
-			$cmd = sprintf("%s -R %s -H %s", $this->externalCmd, $reference, $hypothesis);
-			exec($cmd, $output, $return);
-			unlink($reference);
-			unlink($hypothesis);
-			if ($return != 0){
-				return -1;
-			} else {
-                file_put_contents($temporaryResultsFile, implode("\n", $output));
-				foreach ($output as $line){
-					$tempArray = explode("\t", $line);
-					$typeName = trim(str_replace(":","",$tempArray[0]));
-					if ($typeName == $this->type){
-						return $tempArray[1]; // if you want to use % representation, use [2] instead
-					}
-				}
-			}
-		}
+            return $this->processOutput($output);
+          }
 		return -1;
 	}
